Changed Hand and Card to use ints for card values because why the fuck not

+++ Poker Hands Comparator/Card.cs
@@ -2,10 +2,10 @@
 {
     class Card
     {
-        private byte value;
+        private int value;
         private char suit;
 
-        public Card(byte _value, char _suit)
+        public Card(int _value, char _suit)
         {
             this.value = _value;
             this.suit = _suit;
